+++ frontend/src/components/admin/cms/CmsAdminsTab.vue
@@ -1,0 +1,13 @@
+<template>
+  <div class="cms-tab-content cms-tab-placeholder">
+    <p class="text-sm text-ink-secondary">
+      {{ t("cms.tabs.adminsPlaceholder") }}
+    </p>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { useI18n } from "vue-i18n";
+
+const { t } = useI18n();
+</script>
