+++ frontend/test/views/admin/CMSView.test.ts
@@ -8,6 +8,7 @@ import { useAuthStore } from "@/stores/auth";
 import { RouteNames } from "@/router/routeNames";
 import { i18n } from "@/i18n";
 import CMSView from "@/views/admin/CMSView.vue";
+import CmsProductionsTab from "@/components/admin/cms/CmsProductionsTab.vue";
 import * as productionsService from "@/services/productions";
 import * as tagsService from "@/services/tags";
 import * as hallsService from "@/services/halls";
@@ -227,7 +228,7 @@ describe("CMSView", () => {
 
   it("validates create modal fields before calling createProduction", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     api.openCreateModal();
     await api.submitCreateProduction();
@@ -238,7 +239,7 @@ describe("CMSView", () => {
 
   it("submits create production with valid form data", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     api.openCreateModal();
     api.createForm.value.title.nl = "Title";
@@ -255,7 +256,7 @@ describe("CMSView", () => {
 
   it("opens editor panel when clicking long-text cells and saves", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
     const row = api.rowData.value[0];
 
     api.onCellClicked({
@@ -271,7 +272,7 @@ describe("CMSView", () => {
 
   it("handles create event modal open/close and no-selection submit", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     api.openCreateEventModal();
     expect(api.createEventModalOpen.value).toBe(true);
@@ -285,7 +286,7 @@ describe("CMSView", () => {
 
   it("closes event panel and ignores unrelated cell edit events", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     api.selectedEventsProductionId.value = 42;
     api.selectedEventRows.value = [{ id: 1 }];
@@ -301,7 +302,7 @@ describe("CMSView", () => {
 
   it("covers helper mappers and reset helpers", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     expect(api.localizeValue(undefined)).toBe("");
     expect(api.setCurrentLanguageValue({ en: "old" }, "new")).toMatchObject({ nl: "new" });
@@ -321,7 +322,7 @@ describe("CMSView", () => {
 
   it("covers createAndLinkEvent validation and success branches", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     api.selectedEventsProductionId.value = mockProduction.id;
     api.createLinkedEventForm.value.hallId = 0;
@@ -339,7 +340,7 @@ describe("CMSView", () => {
 
   it("covers save/remove linked event branches", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     const eventRow = {
       id: 100,
@@ -364,7 +365,7 @@ describe("CMSView", () => {
 
   it("covers click/edit handlers and event-row focus behavior", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
     const row = api.rowData.value[0];
 
     api.onCellClicked({ data: row, colDef: { colId: "eventsAction" } });
@@ -394,7 +395,7 @@ describe("CMSView", () => {
 
   it("covers showEventsForProduction success and error branches", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
     const row = api.rowData.value[0];
 
     row.source.events = [100];
@@ -410,7 +411,7 @@ describe("CMSView", () => {
 
   it("covers refreshEventsPanelForSelectedProduction missing-row branch", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     api.selectedEventsProductionId.value = 9999;
     await api.refreshEventsPanelForSelectedProduction();
@@ -420,7 +421,7 @@ describe("CMSView", () => {
 
   it("covers onCellEditingStopped non-inline and revert branches", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
     const row = api.rowData.value[0];
     const setDataValue = vi.fn();
 
@@ -445,7 +446,7 @@ describe("CMSView", () => {
 
   it("covers committed inline edit no-change and error branches", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
     const row = api.rowData.value[0];
     const setDataValue = vi.fn();
 
@@ -473,7 +474,7 @@ describe("CMSView", () => {
 
   it("covers helper branches for language checks and loadCmsData direct call", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     expect(hasAnyLanguageValue({ nl: "", en: "", fr: "" })).toBe(false);
     expect(hasAnyLanguageValue({ nl: "x", en: "", fr: "" })).toBe(true);
@@ -484,7 +485,7 @@ describe("CMSView", () => {
 
   it("covers template interaction handlers and key/file branches", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     const addButton = wrapper.find(".cms-add-button");
     await addButton.trigger("click");
@@ -577,7 +578,7 @@ describe("CMSView", () => {
 
   it("covers additional CMS error and utility branches", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     vi.spyOn(eventsService, "createEvent").mockRejectedValueOnce(new Error("create fail"));
     api.selectedEventsProductionId.value = mockProduction.id;
@@ -630,7 +631,7 @@ describe("CMSView", () => {
 
   it("keeps create-event modal open when submitCreateEvent hits validation error", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     api.selectedEventsProductionId.value = mockProduction.id;
     api.createEventModalOpen.value = true;
@@ -670,7 +671,7 @@ describe("CMSView", () => {
       await cancelBtn.trigger("click");
     }
 
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
     api.selectedEventsProductionId.value = mockProduction.id;
     api.createEventModalOpen.value = true;
     await nextTick();
@@ -693,7 +694,7 @@ describe("CMSView", () => {
 
   it("covers column chooser checkbox, editor textarea model, locale watch and unmount", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
 
     await wrapper.findAll(".cms-grid-actions .cms-mini-btn").at(-1)?.trigger("click");
     const chooserCheckbox = wrapper.find(".cms-column-chooser input[type='checkbox']");
@@ -721,7 +722,7 @@ describe("CMSView", () => {
 
   it("covers remaining event-panel branch paths", async () => {
     const wrapper = await mountCMSView();
-    const api = (wrapper.vm as any).__test;
+    const api = (wrapper.findComponent(CmsProductionsTab).vm as any).__test;
     const row = api.rowData.value[0];
 
     row.source.events = [];
